+++ Courier/Changes.md
@@ -24,6 +24,7 @@ Folder-targeting
 - UTF-8 encoded email addresses are now automatically decoded
 - Attachment names are now normalized before saving them on the disk
 - EmailBuilder now uses ISO-8859 as backup decoding
+- Added implicit conversion from a **Vector** of **Strings** to **Recipients**
 
 ## v1.0.7 - 01.05.2023
 Rebuild due to the changes in other modules.